дз к лекции 7
